write modifications regarding gpl-3.0

+++ contracts/tokens/IStablecoin.sol
@@ -1,3 +1,5 @@
+// SPDX-License-Identifier: Apache-2.0
+
 pragma solidity ^0.8.0;
 
 interface IStablecoin {
@@ -7,5 +9,13 @@ interface IStablecoin {
 
     function burnFrom(address account, uint256 amount) external;
 
-    function permit(address owner, address spender, uint value, uint deadline, uint8 v, bytes32 r, bytes32 s) external;
+    function permit(
+        address owner,
+        address spender,
+        uint256 value,
+        uint256 deadline,
+        uint8 v,
+        bytes32 r,
+        bytes32 s
+    ) external;
 }
